chore: restructure WidgetsEditor (#31803)
## Description
In this PR we are restructuring the WidgetsEditor into multiple
components so that it is easy to assess changes.
currently the widgets editor just does too much at one place to even
understand what is happening.

This is also the first step in trying to decrease the number of places
where layout system type is being used to manipulate a functionality.

Fixes #`Issue Number`  
_or_  
Fixes `Issue URL`
> [!WARNING]  
> _If no issue exists, please create an issue first, and check with the
maintainers if the issue is valid._

## Automation

/ok-to-test tags="@tag.All"

### :mag: Cypress test results
