+++ app/client/src/pages/Editor/WidgetsEditor/components/PropertyPaneWrapper.tsx
@@ -3,10 +3,13 @@ import PropertyPaneSidebar from "components/editorComponents/PropertyPaneSidebar
 import React, { useCallback } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { getPropertyPaneWidth } from "selectors/propertyPaneSelectors";
-import { CreateNewModal } from "../IDE/RightPane/components/CreateNewModal";
+import { CreateNewModal } from "pages/Editor/IDE/RightPane/components/CreateNewModal";
 
 /**
- * OldName: PropertyPaneContainer
+ * PropertyPaneWrapper
+ *
+ * This component is used to wrap the property pane sidebar and create new modal.
+ * It is used to handle the width of the property pane sidebar.
  */
 function PropertyPaneWrapper() {
   const dispatch = useDispatch();
